# contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ampara Animal — contrato do Ciclo 1",
    "version": "0.3.0",
    "description": "Contrato do Ciclo 1 (Sprint 3) e do prontuário do Ciclo 2. Endpoints marcados x-situacao=implementado existem no código e são conferidos por scripts/validar-openapi.cjs. Todas as rotas exigem login (Authorization: Bearer <access token do Supabase Auth>), exceto /api/health e /api/health/ready. IDs e dados dos exemplos são sintéticos. Autorização: além do login, a conta precisa estar liberada na API (tabela Usuario); conta não liberada ou desativada recebe 403. ADMIN acessa todas as unidades; OPERADOR só as vinculadas (papéis provisórios até aprovação da Ampara). O escopo é decidido no servidor: filtros enviados pelo cliente nunca ampliam o acesso."
  },
  "paths": {
...
    "/api/v1/animals": {
      "get": {
        "summary": "Listar animais ativos",
        "description": "Implementado. Lista animais ativos (não arquivados), mais recentes primeiro, com nomes de espécie/unidade/localização e URL assinada da foto (gerada em lote). Filtros combináveis; texto vazio é ignorado. Restrito às unidades do usuário: o filtro unidadeId não amplia o acesso.",
        "x-situacao": "implementado",
        "security": [
          {
            "bearerSupabase": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado esperado quando implementado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaAnimais"
                }
              }
            }
          },
          "400": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Sem token, token inválido ou vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "403": {
            "description": "Conta não liberada/desativada, ou registro fora das unidades do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "422": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "429": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "500": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "503": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "pagina",
            "in": "query",
            "required": false,
            "description": "Filtro ou controle da requisição",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "description": "Filtro ou controle da requisição",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "busca",
            "in": "query",
            "required": false,
            "description": "Busca livre: nome (contém), identificador público (contém) ou microchip (começa com), sem diferenciar maiúsculas",
            "schema": {
              "type": "string",
              "maxLength": 120
            }
          },
          {
            "name": "nome",
            "in": "query",
            "required": false,
            "description": "Nome contém, sem diferenciar maiúsculas",
            "schema": {
              "type": "string",
              "maxLength": 120
            }
          },
          {
            "name": "identificadorPublico",
            "in": "query",
            "required": false,
            "description": "Identificador público exato, sem diferenciar maiúsculas",
            "schema": {
              "type": "string",
              "maxLength": 40
            }
          },
          {
            "name": "microchip",
            "in": "query",
            "required": false,
            "description": "Microchip exato",
            "schema": {
              "type": "string",
              "pattern": "^[0-9]{1,20}$"
            }
          },
          {
            "name": "especieId",
            "in": "query",
            "required": false,
            "description": "Espécie",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "frente",
            "in": "query",
            "required": false,
            "description": "Frente atual",
            "schema": {
              "type": "string",
              "enum": [
                "CCPA",
                "CASADOTE",
                "CED"
              ]
            }
          },
          {
            "name": "unidadeId",
            "in": "query",
            "required": false,
            "description": "Unidade atual",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      },
      "post": {
        "summary": "Concluir admissão com foto confirmada",
        "description": "Implementado. Exige foto confirmada e aplica as regras por frente (ver CriarAnimal). 409: microchip já cadastrado em animal ativo (detalhes.animalExistenteId) ou foto já vinculada.",
        "x-situacao": "implementado",
        "security": [
          {
            "bearerSupabase": []
          }
        ],
        "responses": {
          "201": {
            "description": "Animal cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Animal"
                }
              }
            }
          },
          "400": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Sem token, token inválido ou vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "403": {
            "description": "Conta não liberada/desativada, ou registro fora das unidades do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "422": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "429": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "500": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "503": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarAnimal"
              }
            }
          }
        }
      }
    },
    "/api/v1/animals/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Identificador do recurso",
          "schema": {
            "type": "string",
            "description": "Identificador UUID",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "summary": "Consultar animal",
        "description": "Implementado. Retorna o animal se ele existir e não estiver arquivado.",
        "x-situacao": "implementado",
        "security": [
          {
            "bearerSupabase": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado esperado quando implementado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Animal"
                }
              }
            }
          },
          "400": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Sem token, token inválido ou vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "403": {
            "description": "Conta não liberada/desativada, ou registro fora das unidades do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "422": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "429": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "500": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "503": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Atualizar animal verificando versão",
        "description": "Implementado. Regras avaliadas no estado final; 409 para versão desatualizada ou microchip de outro animal.",
        "x-situacao": "implementado",
        "security": [
          {
            "bearerSupabase": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado esperado quando implementado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Animal"
                }
              }
            }
          },
          "400": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Sem token, token inválido ou vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "403": {
            "description": "Conta não liberada/desativada, ou registro fora das unidades do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "422": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "429": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "500": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "503": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarAnimal"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Arquivar animal",
        "description": "Implementado. Arquivamento lógico (soft delete); não remove o registro do banco.",
        "x-situacao": "implementado",
        "security": [
          {
            "bearerSupabase": []
          }
        ],
        "responses": {
          "204": {
            "description": "Operação concluída sem corpo"
          },
          "400": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Sem token, token inválido ou vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "403": {
            "description": "Conta não liberada/desativada, ou registro fora das unidades do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "422": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "429": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "500": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "503": {
            "description": "Erro padronizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Erro": {
        "type": "object",
        "required": [
          "statusCode",
          "mensagem",
          "caminho",
          "timestamp"
        ],
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "mensagem": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "caminho": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "description": "Instante ISO-8601",
            "format": "date-time"
          },
          "detalhes": {
            "type": "object",
            "additionalProperties": true,
            "description": "Opcional, só em alguns 4xx. Ex.: 409 de microchip duplicado traz animalExistenteId e identificadorPublico para o frontend oferecer a transferência."
          }
        },
        "example": {
          "statusCode": 409,
          "mensagem": "Versão desatualizada",
          "caminho": "/api/v1/animals/56b9f220-3c58-4adc-9dfa-102b90720e12",
          "timestamp": "2026-09-13T12:00:00Z"
        }
      },
...
      "CriarAnimal": {
        "type": "object",
        "additionalProperties": false,
        "description": "Dados de admissão. O identificador público é gerado pelo servidor; foto de entrada confirmada é obrigatória. Regras por frente (validadas no servidor): CCPA e CED exigem microchip, sexo, idadeAproximadaMeses, pesoKg e porte; CASADOTE não exige campos adicionais; espécies silvestres exigem nome e dispensam os demais.",
        "required": [
          "especieId",
          "unidadeId",
          "localizacaoId",
          "responsavelId",
          "frente",
          "fotoEntradaId"
        ],
        "properties": {
          "nome": {
            "type": "string",
            "description": "Nome conhecido. Obrigatório para espécies silvestres (identificação por nome).",
            "nullable": true,
            "maxLength": 120
          },
          "microchip": {
            "type": "string",
            "description": "Somente dígitos, guardado como texto para preservar zeros. Obrigatório nas frentes CCPA e CED (exceto silvestres). Único entre animais ativos: duplicado retorna 409 com o animal existente em detalhes.",
            "example": "000123456789",
            "nullable": true,
            "pattern": "^[0-9]{1,20}$"
          },
          "especieId": {
            "type": "string",
            "description": "Identificador UUID",
            "format": "uuid"
          },
          "racaId": {
            "type": "string",
            "description": "Identificador UUID",
            "format": "uuid",
            "nullable": true
          },
          "unidadeId": {
            "type": "string",
            "description": "Identificador UUID",
            "format": "uuid"
          },
          "localizacaoId": {
            "type": "string",
            "description": "Identificador UUID",
            "format": "uuid"
          },
          "responsavelId": {
            "type": "string",
            "description": "Identificador UUID",
            "format": "uuid"
          },
          "frente": {
            "type": "string",
            "description": "Frente atual. CCPA e CED são frentes de entrada (exigem identificação completa); CASADOTE recebe animais vindos delas.",
            "enum": [
              "CCPA",
              "CASADOTE",
              "CED"
            ]
          },
          "dataEntrada": {
            "type": "string",
            "description": "Instante ISO-8601",
            "format": "date-time",
            "nullable": true
          },
          "sexo": {
            "type": "string",
            "description": "Obrigatório nas frentes CCPA e CED (exceto silvestres).",
            "enum": [
              "MACHO",
              "FEMEA"
            ],
            "nullable": true
          },
          "idadeAproximadaMeses": {
            "type": "integer",
            "minimum": 0,
            "nullable": true,
            "description": "Meses. Obrigatório nas frentes CCPA e CED (exceto silvestres).",
            "maximum": 600
          },
          "pesoKg": {
            "type": "number",
            "minimum": 0,
            "nullable": true,
            "description": "Peso atual em kg, até duas casas decimais; sem histórico. Obrigatório nas frentes CCPA e CED (exceto silvestres).",
            "maximum": 9999.99,
            "multipleOf": 0.01
          },
          "porte": {
            "type": "string",
            "description": "Obrigatório nas frentes CCPA e CED (exceto silvestres).",
            "enum": [
              "PEQUENO",
              "MEDIO",
              "GRANDE"
            ],
            "nullable": true
          },
          "cor": {
            "type": "string",
            "description": "Cor",
            "nullable": true,
            "maxLength": 60
          },
          "observacoes": {
            "type": "string",
            "description": "Ex.: mancha branca na pata dianteira",
            "nullable": true,
            "maxLength": 2000
          },
          "fotoEntradaId": {
            "type": "string",
            "format": "uuid",
            "nullable": false,
            "description": "Foto de entrada já confirmada (POST /fotos e POST /fotos/{id}/confirmacao)."
          }
        },
        "example": {
          "nome": "Luna",
          "microchip": null,
          "especieId": "56b9f220-3c58-4adc-9dfa-102b90720e12",
          "racaId": null,
          "unidadeId": "56b9f220-3c58-4adc-9dfa-102b90720e12",
          "localizacaoId": "56b9f220-3c58-4adc-9dfa-102b90720e12",
          "responsavelId": "56b9f220-3c58-4adc-9dfa-102b90720e12",
          "frente": "CASADOTE",
          "observacoes": null,
          "fotoEntradaId": "56b9f220-3c58-4adc-9dfa-102b90720e12"
        }
      },
...
      "Animal": {
        "type": "object",
        "required": [
          "id",
          "identificadorPublico",
          "versao",
          "criadoEm",
          "atualizadoEm"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Identificador UUID",
            "format": "uuid"
          },
          "identificadorPublico": {
            "type": "string",
            "description": "Identificador público gerado pelo servidor (AM-AAAA-XXXXXXXX)"
          },
          "nome": {
            "type": "string",
            "description": "Nome conhecido",
            "nullable": true
          },
          "microchip": {
            "type": "string",
            "description": "Somente dígitos; único entre animais ativos",
            "example": "000123456789",
            "nullable": true
          },
          "especieId": {
            "type": "string",
            "description": "Identificador UUID",
            "format": "uuid"
          },
          "racaId": {
            "type": "string",
            "description": "Identificador UUID",
            "format": "uuid",
            "nullable": true
          },
          "unidadeId": {
            "type": "string",
            "description": "Identificador UUID",
            "format": "uuid"
          },
          "localizacaoId": {
            "type": "string",
            "description": "Identificador UUID",
            "format": "uuid"
          },
          "responsavelId": {
            "type": "string",
            "description": "Identificador UUID",
            "format": "uuid"
          },
          "frente": {
            "type": "string",
            "description": "Classificação atual; demais nomes aguardam validação",
            "enum": [
              "CCPA",
              "CASADOTE",
              "CED"
            ]
          },
          "dataEntrada": {
            "type": "string",
            "description": "Instante ISO-8601",
            "format": "date-time",
            "nullable": true
          },
          "sexo": {
            "type": "string",
            "description": "Sexo informado",
            "enum": [
              "MACHO",
              "FEMEA"
            ],
            "nullable": true
          },
          "idadeAproximadaMeses": {
            "type": "integer",
            "minimum": 0,
            "nullable": true,
            "description": "Meses; unidade proposta"
          },
          "pesoKg": {
            "type": "number",
            "minimum": 0,
            "nullable": true,
            "description": "Peso atual em kg, duas casas decimais; não há histórico"
          },
          "porte": {
            "type": "string",
            "description": "Porte",
            "enum": [
              "PEQUENO",
              "MEDIO",
              "GRANDE"
            ],
            "nullable": true
          },
          "cor": {
            "type": "string",
            "description": "Cor",
            "nullable": true
          },
          "versao": {
            "type": "integer",
            "minimum": 1
          },
          "criadoEm": {
            "type": "string",
            "description": "Instante ISO-8601",
            "format": "date-time"
          },
          "atualizadoEm": {
            "type": "string",
            "description": "Instante ISO-8601",
            "format": "date-time"
          },
          "arquivadoEm": {
            "type": "string",
            "description": "Instante ISO-8601",
            "format": "date-time",
            "nullable": true
          },
          "observacoes": {
            "type": "string",
            "description": "Ex.: mancha branca na pata dianteira",
            "nullable": true
          },
          "fotoEntradaId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Identificador UUID"
          },
          "especieNome": {
            "type": "string",
            "description": "Nome da espécie (evita consulta extra no frontend)"
          },
          "especieSilvestre": {
            "type": "boolean",
            "description": "Espécie silvestre (identificação por nome)"
          },
          "unidadeNome": {
            "type": "string"
          },
          "localizacaoNome": {
            "type": "string"
          },
          "fotoEntradaUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "description": "URL assinada de leitura da foto de entrada, válida por 1 hora. Null se o armazenamento estiver indisponível."
          }
        },
        "example": {
          "id": "56b9f220-3c58-4adc-9dfa-102b90720e12",
          "identificadorPublico": "EXEMPLO-001",
          "nome": "Luna",
          "microchip": null,
          "especieId": "56b9f220-3c58-4adc-9dfa-102b90720e12",
          "racaId": null,
          "unidadeId": "56b9f220-3c58-4adc-9dfa-102b90720e12",
          "localizacaoId": "56b9f220-3c58-4adc-9dfa-102b90720e12",
          "responsavelId": "56b9f220-3c58-4adc-9dfa-102b90720e12",
          "frente": "CASADOTE",
          "fotoEntradaId": null,
          "versao": 1,
          "criadoEm": "2026-09-13T12:00:00Z",
          "atualizadoEm": "2026-09-13T12:00:00Z",
          "arquivadoEm": null,
          "observacoes": null
        }
      },
...
      "PaginaAnimais": {
        "type": "object",
        "required": [
          "itens",
          "pagina",
          "limite",
          "total"
        ],
        "properties": {
          "itens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Animal"
            }
          },
          "pagina": {
            "type": "integer",
            "minimum": 1
          },
          "limite": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerSupabase": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token emitido pelo Supabase Auth do projeto (o mesmo login do frontend). Validado no servidor pela assinatura (JWKS), emissor e audiência \"authenticated\". Sem token ou com token inválido/vencido: 401."
      }
    }
  }
}