# src/animals/animals.controller.ts
import {
  Body,
  Controller,
  Delete,
  Get,
  HttpCode,
  HttpStatus,
  Param,
  ParseUUIDPipe,
  Patch,
  Post,
  Query,
} from '@nestjs/common';
import { ApiOperation, ApiTags } from '@nestjs/swagger';
import { EscopoAtual } from '../auth/decoradores';
import { EscopoAcesso } from '../auth/escopo';
import { AnimalsService } from './animals.service';
import { AtualizarAnimalDto } from './dto/atualizar-animal.dto';
import { CriarAnimalDto } from './dto/criar-animal.dto';
import { ListarAnimaisQueryDto } from './dto/listar-animais-query.dto';

@ApiTags('Animals')
@Controller({ path: 'animals', version: '1' })
export class AnimalsController {
  constructor(private readonly animalsService: AnimalsService) {}

  @Get()
  @ApiOperation({ summary: 'Listar animais ativos' })
  listar(@Query() query: ListarAnimaisQueryDto, @EscopoAtual() escopo: EscopoAcesso) {
    return this.animalsService.listar(query, escopo);
  }

  @Post()
  @HttpCode(HttpStatus.CREATED)
  @ApiOperation({ summary: 'Cadastrar animal (admissão)' })
  criar(@Body() dto: CriarAnimalDto, @EscopoAtual() escopo: EscopoAcesso) {
    return this.animalsService.criar(dto, escopo);
  }

  @Get(':id')
  @ApiOperation({ summary: 'Consultar animal' })
  buscarPorId(@Param('id', ParseUUIDPipe) id: string, @EscopoAtual() escopo: EscopoAcesso) {
    return this.animalsService.buscarPorId(id, escopo);
  }

  @Patch(':id')
  @ApiOperation({ summary: 'Atualizar animal verificando versão' })
  atualizar(
    @Param('id', ParseUUIDPipe) id: string,
    @Body() dto: AtualizarAnimalDto,
    @EscopoAtual() escopo: EscopoAcesso,
  ) {
    return this.animalsService.atualizar(id, dto, escopo);
  }

  @Delete(':id')
  @HttpCode(HttpStatus.NO_CONTENT)
  @ApiOperation({ summary: 'Arquivar animal' })
  arquivar(@Param('id', ParseUUIDPipe) id: string, @EscopoAtual() escopo: EscopoAcesso) {
    return this.animalsService.arquivar(id, escopo);
  }
}